`test`: Static-Adapter bandaid fix
So I've been dealing with trying to get an image loaded correctly in the build environment. The `build` folder correctly contains all assets from `static` which indicates that the `static-adapter` works fine. The issue has been that actually trying to load the image fails due to incorrect path.I followed the `static-adapter`'s README instructions and added

```js
paths: {
       base: dev ? '' : '/website',
     },
```
logic to set the `base` variable based on which environemnt we are in. This has been the issue all along, though. Leaving the `base` variable unchanged allows the image in `+page.svelte` to load correctly in my local test. I am now testing whether it still works correctly in `gh-pages`.